refactoring to stay consistent with stories
